Refactoring approve accommodations. Adding check/exam update for custom accommodations
