+++ service/src/main/java/tds/exam/services/ExamAccommodationService.java
@@ -3,14 +3,13 @@ package tds.exam.services;
 import org.joda.time.Instant;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 import tds.assessment.Assessment;
-import tds.common.ValidationError;
 import tds.exam.ApproveAccommodationsRequest;
 import tds.exam.Exam;
 import tds.exam.ExamAccommodation;
+import tds.session.Session;
 
 /**
  * Handles interaction with {@link tds.exam.ExamAccommodation}s associated to an {@link tds.exam.Exam}
@@ -37,7 +36,7 @@ public interface ExamAccommodationService {
     /**
      * Initializes and inserts exam accommodations for the exam;
      *
-     * @param exam exam to use to initialize the {@link tds.exam.ExamAccommodation}
+     * @param exam                      exam to use to initialize the {@link tds.exam.ExamAccommodation}
      * @param studentAccommodationCodes the separated list of accommodations from the student package retrieved from ART
      */
     List<ExamAccommodation> initializeExamAccommodations(Exam exam, String studentAccommodationCodes);
@@ -53,29 +52,29 @@ public interface ExamAccommodationService {
     /**
      * Initializes accommodations on a previous exam
      *
-     * @param exam                {@link tds.exam.Exam}
-     * @param assessment          {@link tds.assessment.Assessment}
-     * @param segmentPosition     the segment position for the accommodations
-     * @param restoreRts          {@code true} if the restore rts
+     * @param exam                      {@link tds.exam.Exam}
+     * @param assessment                {@link tds.assessment.Assessment}
+     * @param segmentPosition           the segment position for the accommodations
+     * @param restoreRts                {@code true} if the restore rts
      * @param studentAccommodationCodes the student accommodations String
-     *
      * @return list of {@link tds.exam.ExamAccommodation}
      */
     List<ExamAccommodation> initializeAccommodationsOnPreviousExam(Exam exam, Assessment assessment, int segmentPosition, boolean restoreRts, String studentAccommodationCodes);
-    
+
     /**
      * Approves {@link tds.exam.ExamAccommodation}s for an exam and set of accommodation codes.
      *
-     * @param examId  the id of the {@link tds.exam.Exam} to approve accommodations for
+     * @param exam    the {@link tds.exam.Exam} to approve accommodations for
+     * @param session the {@link tds.session.Session} the exam belongs to
      * @param request the {@link tds.exam.ApproveAccommodationsRequest} containing request datazz
-     * @return an optional {@link tds.common.ValidationError}
+     * @return the updated {@link List<tds.exam.ExamAccommodation>}s
      */
-    Optional<ValidationError> approveAccommodations(UUID examId, ApproveAccommodationsRequest request);
+    List<ExamAccommodation> approveAccommodations(final Exam exam, final Session session, final ApproveAccommodationsRequest request);
 
     /**
      * Denies all {@link tds.exam.ExamAccommodation}s for an exam
      *
-     * @param examId The id of the exam for which to deny accommodations
+     * @param examId   The id of the exam for which to deny accommodations
      * @param deniedAt
      */
     void denyAccommodations(final UUID examId, final Instant deniedAt);
