+++ service/src/main/java/tds/exam/web/endpoints/ExamAccommodationController.java
@@ -52,20 +52,4 @@ public class ExamAccommodationController {
     ResponseEntity<List<ExamAccommodation>> findApprovedAccommodations(@PathVariable final UUID examId) {
         return ResponseEntity.ok(examAccommodationService.findApprovedAccommodations(examId));
     }
-    
-    @RequestMapping(value = "/{examId}/accommodations", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
-    ResponseEntity<NoContentResponseResource> approveAccommodations(@PathVariable final UUID examId, @RequestBody ApproveAccommodationsRequest request) {
-        Optional<ValidationError> maybeError = examAccommodationService.approveAccommodations(examId, request);
-        
-        if (maybeError.isPresent()) {
-            NoContentResponseResource response = new NoContentResponseResource(maybeError.get());
-            return new ResponseEntity<>(response, HttpStatus.UNPROCESSABLE_ENTITY);
-        }
-        
-        Link link = linkTo(methodOn(ExamAccommodationController.class).findAccommodations(examId)).withSelfRel();
-        final HttpHeaders headers = new HttpHeaders();
-        headers.add("Location", link.getHref());
-        
-        return new ResponseEntity<>(headers, HttpStatus.NO_CONTENT);
-    }
 }
